Move QueryParsingException to query_framework
