+++ InformationScripting/src/query_framework/QueryParsingException.cpp
