+++ InformationScripting/src/query_framework/QueryParsingException.h
